fix: add support for Relay contract upgrade, VoterRegistry abi upgrade on some networks

+++ common/fsp/voters.go
@@ -59,7 +59,7 @@ func GetVoterIndex(db *gorm.DB, epoch ty.RewardEpochId, fromSec, toSec uint64, p
 
 	var voterInfos []*VoterInfo
 	for _, reg := range regs {
-		if reg.RewardEpochId != uint32(epoch) {
+		if reg.RewardEpochId != uint64(epoch) {
 			continue
 		}
